+++ globus_search_cli/commands/ingest.py
@@ -32,7 +32,7 @@ from globus_search_cli.printing import format_output
     type=int,
     default=100,
     show_default=True,
-    help=("When source-type=GMetaList, how many entries to send " "at a time"),
+    help="When source-type=gmetalist, how many entries to send at a time",
 )
 @index_argument
 @click.argument("documents", nargs=-1, type=click.File())
@@ -54,10 +54,7 @@ def ingest_func(index_id, source_type, batch_size, documents):
 
                 current_doc["gmeta"] = current_slice
                 gingest_doc = {
-                    "@datatype": "GIngest",
-                    "@version": "2016-11-09",
                     "ingest_type": "GMetaList",
-                    "source_id": "search_client_command",
                     "ingest_data": current_doc,
                 }
                 format_output("Sending batch of size {}".format(batch_size))